requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
torch>=2.9.1
torchvision>=0.16.2
pillow>=12.0.0
streamlit>=1.26.0
streamlit-drawable-canvas==0.9.3